+++ core/src/main/java/org/apache/carbondata/core/indexstore/BlockMetaInfo.java
@@ -1,0 +1,47 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.carbondata.core.indexstore;
+
+/**
+ * Holds the metadata info of the block.
+ */
+public class BlockMetaInfo {
+
+  /**
+   * HDFS locations of a block
+   */
+  private String[] locationInfo;
+
+  /**
+   * Size of block
+   */
+  private long size;
+
+  public BlockMetaInfo(String[] locationInfo, long size) {
+    this.locationInfo = locationInfo;
+    this.size = size;
+  }
+
+  public String[] getLocationInfo() {
+    return locationInfo;
+  }
+
+  public long getSize() {
+    return size;
+  }
+}
+
