+++ core/src/main/java/org/apache/carbondata/core/indexstore/blockletindex/BlockletDataMapModel.java
@@ -20,7 +20,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.carbondata.core.datamap.dev.DataMapModel;
-import org.apache.carbondata.core.indexstore.BlockletDataMapIndexStore;
+import org.apache.carbondata.core.indexstore.BlockMetaInfo;
 
 /**
  * It is the model object to keep the information to build or initialize BlockletDataMap.
@@ -33,11 +33,11 @@ public class BlockletDataMapModel extends DataMapModel {
 
   private boolean partitionedSegment;
 
-  Map<String, BlockletDataMapIndexStore.BlockMetaInfo> blockMetaInfoMap;
+  Map<String, BlockMetaInfo> blockMetaInfoMap;
 
   public BlockletDataMapModel(String filePath, byte[] fileData, List<String> partitions,
       boolean partitionedSegment,
-      Map<String, BlockletDataMapIndexStore.BlockMetaInfo> blockMetaInfoMap) {
+      Map<String, BlockMetaInfo> blockMetaInfoMap) {
     super(filePath);
     this.fileData = fileData;
     this.partitions = partitions;
@@ -57,7 +57,7 @@ public class BlockletDataMapModel extends DataMapModel {
     return partitionedSegment;
   }
 
-  public Map<String, BlockletDataMapIndexStore.BlockMetaInfo> getBlockMetaInfoMap() {
+  public Map<String, BlockMetaInfo> getBlockMetaInfoMap() {
     return blockMetaInfoMap;
   }
 }
